requirements.txt
Django==1.9.2
django-hitcount==1.2.1
django-mathjax==0.0.5
django-pagedown==0.1.1
django-registration==2.0.4
django-split-settings==0.1.3
django-taggit==0.18.0
django-vote==1.1.2
hoedown==0.2.2
psycopg2==2.6.1
Pygments==2.1.1
